+++ project-base/src/Shopsys/ShopBundle/Model/Cart/Item/DeleteOldCartsCronModule.php
@@ -2,7 +2,7 @@
 
 namespace Shopsys\ShopBundle\Model\Cart\Item;
 
-use Shopsys\ShopBundle\Component\Cron\SimpleCronModuleInterface;
+use Shopsys\Plugin\Cron\SimpleCronModuleInterface;
 use Shopsys\ShopBundle\Model\Cart\CartFacade;
 use Symfony\Bridge\Monolog\Logger;