+++ project-base/src/Shopsys/ShopBundle/Model/Category/CategoryData.php
@@ -42,11 +42,6 @@ class CategoryData
     public $hiddenOnDomains;
 
     /**
-     * @var \Shopsys\ShopBundle\Model\Feed\Category\FeedCategory|null
-     */
-    public $heurekaCzFeedCategory;
-
-    /**
      * @var \Shopsys\ShopBundle\Form\UrlListData
      */
     public $urls;
@@ -55,6 +50,11 @@ class CategoryData
      * @var string[]
      */
     public $image;
+
+    /**
+     * @var array
+     */
+    public $pluginData;
 
     public function __construct()
     {
@@ -66,6 +66,7 @@ class CategoryData
         $this->hiddenOnDomains = [];
         $this->urls = new UrlListData();
         $this->image = [];
+        $this->pluginData = [];
     }
 
     /**
@@ -81,7 +82,6 @@ class CategoryData
         }
         $this->name = $names;
         $this->parent = $category->getParent();
-        $this->heurekaCzFeedCategory = $category->getHeurekaCzFeedCategory();
         $seoTitles = [];
         $seoMetaDescriptions = [];
         $seoH1 = [];