+++ project-base/src/Shopsys/ShopBundle/Model/Feed/DeliveryFeedCronModule.php
@@ -2,7 +2,7 @@
 
 namespace Shopsys\ShopBundle\Model\Feed;
 
-use Shopsys\ShopBundle\Component\Cron\SimpleCronModuleInterface;
+use Shopsys\Plugin\Cron\SimpleCronModuleInterface;
 use Symfony\Bridge\Monolog\Logger;
 
 class DeliveryFeedCronModule implements SimpleCronModuleInterface